pyproject.toml
[project]
name = "chatbot-langchain-streamlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-google-genai>=2.1.9",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.0",
]
